refactor: reorganize blog components
