+++ src/components/blog/Author.tsx
@@ -1,4 +1,4 @@
-import { AuthorContent } from "../lib/authors";
+import { AuthorContent } from "../../lib/authors";
 
 type Props = {
   author: AuthorContent;
