+++ src/components/blog/Layout.tsx
@@ -1,5 +1,5 @@
 import Head from "next/head";
-import Navigation from "./Navigation";
+import NavBar from "../homepage/navbar";
 
 type Props = {
   children: React.ReactNode;
@@ -14,9 +14,10 @@ export default function Layout({ children }: Props) {
         <link rel="apple-touch-icon" href="/icon.png" />
         <meta name="theme-color" content="#fff" />
       </Head>
-      <nav>
-        <Navigation />
-      </nav>
+      <NavBar />
+      {/* <nav>
+       <Navigation />
+      </nav> */}
       <main>{children}</main>
       <style jsx>
         {`
