+++ src/components/blog/Pagination.tsx
@@ -1,4 +1,4 @@
-import { generatePagination } from "../lib/pagination";
+import { generatePagination } from "../../lib/pagination";
 import Link from "next/link";
 
 type Props = {
