+++ src/components/blog/PostItem.tsx
@@ -1,4 +1,4 @@
-import { PostContent } from "../lib/posts";
+import { PostContent } from "../../lib/posts";
 import Date from "./Date";
 import Link from "next/link";
 import { parseISO } from "date-fns";
