+++ src/components/blog/PostList.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { PostContent } from "../lib/posts";
+import { PostContent } from "../../lib/posts";
 import PostItem from "./PostItem";
 import TagLink from "./TagLink";
 import Pagination from "./Pagination";
-import { TagContent } from "../lib/tags";
+import { TagContent } from "../../lib/tags";
 
 type Props = {
   posts: PostContent[];
