+++ src/components/blog/TagLink.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { TagContent } from "../lib/tags";
+import { TagContent } from "../../lib/tags";
 
 type Props = {
   tag: TagContent;
