+++ src/components/blog/TagPostList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { PostContent } from "../lib/posts";
-import { TagContent } from "../lib/tags";
+import { PostContent } from "../../lib/posts";
+import { TagContent } from "../../lib/tags";
 import PostItem from "./PostItem";
 import Pagination from "./Pagination";
