+++ src/components/blog/meta/JsonLdMeta.tsx
@@ -1,6 +1,6 @@
 import { BlogPosting } from "schema-dts";
 import { jsonLdScriptProps } from "react-schemaorg";
-import config from "../../lib/config";
+import config from "../../../lib/config";
 import { formatISO } from "date-fns";
 import Head from "next/head";
