+++ src/pages/blog/[post].tsx
@@ -1,13 +1,12 @@
 import { GetStaticProps, GetStaticPaths } from "next";
 import { serialize } from "next-mdx-remote/serialize";
-// import { MdxRemote } from "next-mdx-remote/types";
 import { MDXRemote, MDXRemoteSerializeResult } from "next-mdx-remote";
 import matter from "gray-matter";
 import { fetchPostContent } from "../../lib/posts";
 import fs from "fs";
 import yaml from "js-yaml";
 import { parseISO } from "date-fns";
-import PostLayout from "../../components/PostLayout";
+import PostLayout from "../../components/blog/PostLayout";
 
 import InstagramEmbed from "react-instagram-embed";
 import YouTube from "react-youtube";
