+++ src/pages/blog/page/[page].tsx
@@ -1,10 +1,10 @@
 import { GetStaticPaths, GetStaticProps } from "next";
 import Head from "next/head";
-import Layout from "../../../components/Layout";
-import BasicMeta from "../../../components/meta/BasicMeta";
-import OpenGraphMeta from "../../../components/meta/OpenGraphMeta";
-import TwitterCardMeta from "../../../components/meta/TwitterCardMeta";
-import PostList from "../../../components/PostList";
+import Layout from "../../../components/blog/Layout";
+import BasicMeta from "../../../components/blog/meta/BasicMeta";
+import OpenGraphMeta from "../../../components/blog/meta/OpenGraphMeta";
+import TwitterCardMeta from "../../../components/blog/meta/TwitterCardMeta";
+import PostList from "../../../components/blog/PostList";
 import config from "../../../lib/config";
 import { countPosts, listPostContent, PostContent } from "../../../lib/posts";
 import { listTags, TagContent } from "../../../lib/tags";
